renamed database helper
