+++ src/server/routes/__tests__/submissions.integration.spec.js
@@ -1,5 +1,5 @@
 import { Submission } from '../../../entities/index.js'
-import { deleteActivitiesAndSubmissions } from '../../../test-utils/database-helpers.js'
+import { deleteActivitiesAndSubmissions } from '../../../test-utils/database-test-utils.js'
 import initialiseServer from '../../server.js'
 
 describe('submissions.integration', () => {
